Move tasks to setup.py
