+++ setup.py
@@ -1,8 +1,12 @@
 #!/usr/bin/env python
 
-import re
 import ast
-from setuptools import setup, find_packages
+import re
+import subprocess
+import sys
+
+from setuptools import Command, find_packages, setup
+from setuptools.command.test import test as TestCommand
 
 _version_re = re.compile(r'__version__\s+=\s+(.*)')
 
@@ -23,6 +27,45 @@ install_requirements = [
     'cli_helpers[styles] >= 1.0.1',
 ]
 
+
+class lint(Command):
+    description = 'check code against PEP 8 (and fix violations)'
+
+    user_options = [
+        ('branch=', 'b', 'branch/revision to compare against (e.g. master)'),
+        ('fix', 'f', 'fix the violations in place')
+    ]
+
+    def initialize_options(self):
+        """Set the default options."""
+        self.branch = 'master'
+        self.fix = False
+
+    def finalize_options(self):
+        pass
+
+    def run(self):
+        cmd = 'pep8radius {} {}'.format(
+            self.branch, '--in-place' if self.fix else '')
+        sys.exit(subprocess.call(cmd, shell=True))
+
+
+class test(TestCommand):
+
+    user_options = [('pytest-args=', 'a', 'Arguments to pass to pytest')]
+
+    def initialize_options(self):
+        TestCommand.initialize_options(self)
+        self.pytest_args = ''
+
+    def run_tests(self):
+        import shlex
+        import pytest
+        unit_test_errno = pytest.main(shlex.split(self.pytest_args))
+        cli_errno = subprocess.call('behave test/features', shell=True)
+        sys.exit(unit_test_errno or cli_errno)
+
+
 setup(
     name='mycli',
     author='Mycli Core Team',
@@ -36,11 +79,8 @@ setup(
     install_requires=install_requirements,
     entry_points={
         'console_scripts': ['mycli = mycli.main:cli'],
-        'distutils.commands': [
-            'lint = tasks:lint',
-            'test = tasks:test',
-        ],
     },
+    cmdclass={'lint': lint, 'test': test},
     classifiers=[
         'Intended Audience :: Developers',
         'License :: OSI Approved :: BSD License',
